Enhance interactivity by increasing data limit and map zoom level
Changes have been made to the 'dashboard2.py' page, including the increase of the data limit fetched from 100 to 1000 records for better real-time visualization. The initial zoom level of the folium.Map feature has also been heightened from 20 to 18, providing a closer and potentially a more interactive view for the users.